All shader code 64-bit variables demoted to 32-bit

+++ modules/base/shaders/arrow_vs.glsl
@@ -36,7 +36,7 @@ uniform mat4 projectionTransform;
 uniform mat3 normalTransform;
 
 void main() {
-  vs_positionViewSpace = vec4(modelViewTransform * dvec4(in_position, 1));
+  vs_positionViewSpace = vec4(modelViewTransform * vec4(in_position, 1));
   vec4 positionScreenSpace = projectionTransform * vs_positionViewSpace;
   vs_depth = positionScreenSpace.w;
   vs_normal = normalize(normalTransform * in_normal);
